Two .txt files were implemented in the code, one with the questions for the bot and another for the answers. In addition, work was done in the run method so that it does the following:
Read a list of questions from the "questions.txt" file.
Send each question to the customer through the output stream.

Receives and manages the customer's choice
Read the customer's choice from the input stream.
Read a list of responses from the "responses.txt" file.
Sends the corresponding response to the client based on their choice, or
a message if the choice is invalid.

close resources
Closes the input and output streams.
Closes the ServerSocket object.
Handle exceptions:

It catches any exception that may occur during the process and prints it
to the console.